+++ resiliosync/rootfs/etc/cont-init.d/00-migrate.sh
@@ -10,7 +10,7 @@ config_location="${config_location/\/config/\/homeassistant}"
 if [ -f "$config_location/config.yaml" ] && [ ! -f "$config_location/migrated" ]; then
     bashio::log.warning "Migrating config.yaml"
     cp -rf "/homeassistant/addons_config/$slug"/* /config/ || true
-    echo "Migrated to internal config folder accessible at /app_configs/xxx-$slug" > "/homeassistant/addons_config/$slug/migrated"
+    echo "Migrated to internal config folder accessible at /addon_configs/xxx-$slug" > "/homeassistant/addons_config/$slug/migrated"
 fi
 
 if [ -f "/homeassistant/addons_autoscripts/$slug.sh" ]; then